+++ test/prototype/moe_training/test_fp8_blockwise_deepgemm_backend.py
@@ -14,11 +14,11 @@ from torch.fx.experimental.symbolic_shapes import ShapeEnv, has_free_symbols
 pytest.importorskip("triton", reason="Triton required for blockwise FP8 modules")
 
 from torchao.float8.config import e4m3_dtype
-from torchao.prototype.blockwise_fp8_training import (
+from torchao.prototype.moe_training.kernels.fp8_blockwise import (
     deepgemm_grouped_kernels,
     grouped_kernels,
 )
-from torchao.prototype.blockwise_fp8_training.deepgemm_metadata import (
+from torchao.prototype.moe_training.kernels.fp8_blockwise.deepgemm_metadata import (
     build_deepgemm_grouped_offset_plan,
     group_sizes_from_offsets,
 )
@@ -240,13 +240,13 @@ def test_deepgemm_grouped_layout_from_padded_offsets():
 )
 @pytest.mark.parametrize("expert_step", [1, 2])
 def test_grouped_weight_quant_layouts_match_dense_per_expert_quantizers(expert_step):
-    from torchao.prototype.blockwise_fp8_training.grouped_weight_quant import (
-        triton_fp8_blockwise_weight_quant_grouped_dgrad_rhs,
-        triton_fp8_blockwise_weight_quant_grouped_forward_rhs,
-    )
     from torchao.prototype.blockwise_fp8_training.kernels import (
         triton_fp8_blockwise_weight_quant_rhs,
         triton_fp8_blockwise_weight_quant_transposed_rhs,
+    )
+    from torchao.prototype.moe_training.kernels.fp8_blockwise.grouped_weight_quant import (
+        triton_fp8_blockwise_weight_quant_grouped_dgrad_rhs,
+        triton_fp8_blockwise_weight_quant_grouped_forward_rhs,
     )
 
     def stack_per_expert_quant(quant_fn, weight):
@@ -320,7 +320,7 @@ def test_grouped_weight_quant_supports_compile_fullgraph(
 ):
     from torch._dynamo.testing import CompileCounterWithBackend
 
-    from torchao.prototype.blockwise_fp8_training.grouped_weight_quant import (
+    from torchao.prototype.moe_training.kernels.fp8_blockwise.grouped_weight_quant import (
         triton_fp8_blockwise_weight_quant_grouped_dgrad_rhs,
         triton_fp8_blockwise_weight_quant_grouped_forward_rhs,
     )
@@ -353,7 +353,7 @@ def test_grouped_weight_quant_supports_compile_fullgraph(
 
 
 def test_deepgemm_k_grouped_activation_quant_fake_contract_tracks_valid_tokens():
-    from torchao.prototype.blockwise_fp8_training.deepgemm_quant import (
+    from torchao.prototype.moe_training.kernels.fp8_blockwise.deepgemm_quant import (
         triton_fp8_blockwise_act_quant_k_grouped_deepgemm,
     )
 
@@ -391,12 +391,12 @@ def test_deepgemm_k_grouped_activation_quant_fake_contract_tracks_valid_tokens()
 def test_deepgemm_k_grouped_activation_quant_matches_flattened_torchao_layouts(
     offsets,
 ):
-    from torchao.prototype.blockwise_fp8_training.deepgemm_quant import (
-        triton_fp8_blockwise_act_quant_k_grouped_deepgemm,
-    )
     from torchao.prototype.blockwise_fp8_training.kernels import (
         triton_fp8_blockwise_act_quant_rhs,
         triton_fp8_blockwise_act_quant_transposed_lhs,
+    )
+    from torchao.prototype.moe_training.kernels.fp8_blockwise.deepgemm_quant import (
+        triton_fp8_blockwise_act_quant_k_grouped_deepgemm,
     )
 
     def flatten_transposed_lhs(q, group_sizes):
