+++ torchao/prototype/moe_training/kernels/fp8_blockwise/__init__.py
@@ -1,0 +1,27 @@
+from torchao.prototype.moe_training.kernels.fp8_blockwise.deepgemm_grouped_kernels import (
+    can_use_deepgemm_grouped_training as can_use_deepgemm_grouped_training,
+)
+from torchao.prototype.moe_training.kernels.fp8_blockwise.deepgemm_grouped_kernels import (
+    deepgemm_blockwise_scaled_grouped_mm as deepgemm_blockwise_scaled_grouped_mm,
+)
+from torchao.prototype.moe_training.kernels.fp8_blockwise.deepgemm_grouped_kernels import (
+    deepgemm_blockwise_scaled_grouped_mm_wgrad as deepgemm_blockwise_scaled_grouped_mm_wgrad,
+)
+from torchao.prototype.moe_training.kernels.fp8_blockwise.deepgemm_grouped_kernels import (
+    prepare_deepgemm_wgrad_plan as prepare_deepgemm_wgrad_plan,
+)
+from torchao.prototype.moe_training.kernels.fp8_blockwise.deepgemm_metadata import (
+    DeepGemmGroupedOffsetPlan as DeepGemmGroupedOffsetPlan,
+)
+from torchao.prototype.moe_training.kernels.fp8_blockwise.deepgemm_metadata import (
+    build_deepgemm_grouped_offset_plan as build_deepgemm_grouped_offset_plan,
+)
+from torchao.prototype.moe_training.kernels.fp8_blockwise.grouped_kernels import (
+    emulated_blockwise_scaled_grouped_mm as emulated_blockwise_scaled_grouped_mm,
+)
+from torchao.prototype.moe_training.kernels.fp8_blockwise.grouped_weight_quant import (
+    triton_fp8_blockwise_weight_quant_grouped_dgrad_rhs as triton_fp8_blockwise_weight_quant_grouped_dgrad_rhs,
+)
+from torchao.prototype.moe_training.kernels.fp8_blockwise.grouped_weight_quant import (
+    triton_fp8_blockwise_weight_quant_grouped_forward_rhs as triton_fp8_blockwise_weight_quant_grouped_forward_rhs,
+)
